+++ app/javascript/applications/CharKeeperApp/components/substances/Equipment.jsx
@@ -81,7 +81,7 @@ const TRANSLATION = {
   }
 }
 const CREATE_HOMEBREW_ITEMS = ['daggerheart', 'dnd2024'];
-const ITEMS_INFO = ['daggerheart'];
+const ITEMS_INFO = ['daggerheart', 'dnd2024', 'dnd5'];
 const HOMEBREWED_PROVIDERS = ['daggerheart', 'dnd2024']
 
 export const Equipment = (props) => {
@@ -246,7 +246,7 @@ export const Equipment = (props) => {
 
     if (result.errors_list === undefined) {
       batch(() => {
-        if (item.kind.includes('weapon')) props.onReloadCharacter();
+        if (props.weaponsKinds && props.weaponsKinds.includes(item.kind) || item.kind.includes('weapon')) props.onReloadCharacter();
         reloadCharacterItems();
         renderNotice(t('alerts.itemIsAdded'));
       });